address round 2 comments
